pyproject.toml
[project]
name = "eda-hospital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.12"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain-openai>=1.0.2",
    "load-dotenv>=0.1.0",
    "paddleocr[doc-parser]>=3.3.1",
    "paddlepaddle-gpu>=3.2.1",
    "safetensors",
    "langchain-chroma>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "pydantic-settings>=2.11.0",
]

[[tool.uv.index]]
name = "paddle-cu130"
url = "https://www.paddlepaddle.org.cn/packages/stable/cu130/"
explicit = true

[tool.uv.sources]
paddlepaddle-gpu = { index = "paddle-cu130" }
safetensors = { url = "https://paddle-whl.bj.bcebos.com/nightly/cu126/safetensors/safetensors-0.6.2.dev0-cp38-abi3-linux_x86_64.whl" }
